+++ docs/examples/createitems/controller.js
@@ -1,0 +1,26 @@
+angular.module('selectDemo')
+    .controller('selectCreateitemsController', function ($q, $timeout, $scope, ShopArr, ShopArrShort) {
+
+        $scope.shopArrShort = ShopArrShort.query();
+        $scope.shopArr = ShopArr.query();
+
+        $scope.bundle1 =['boots'];
+
+        $scope.shopArr.$promise.then(function(data) {
+            $scope.bundle2 = [data[5]];
+            $scope.bundle3 = [data[3]];
+        });
+
+
+        var counter = 0;
+
+        $scope.addItem = function(query) {
+            var id = counter++;
+
+            return {
+                id: id,
+                name: query + '-' + id,
+                category: "shoes"
+            };
+        };
+    });